test: consolidate SuperJSON tests into dbosjson.test.ts
Merged all DBOSJSON tests into single file with clear organization:
• Original tests for dates, buffers, bigints remain unchanged
• Added SuperJSON enhanced types section for new capabilities
• Added backwards compatibility section to verify legacy data works

Cleaner structure, no separate test file needed.
